+++ packages/cubejs-server/examples/simple-mysql/index.js
@@ -1,7 +1,0 @@
-const CubejsServer = require('@cubejs-backend/server');
-
-const server = new CubejsServer();
-
-server.listen().then(({ port }) => {
-  console.log(`Cube.js server listening on ${port}`);
-});